Homepage finished

+++ controllers/UserController.php
@@ -1,7 +1,5 @@
 <?php
 
-//require_once('Entities/User.php');
-
 class UserController extends Controller
 { 
     public function __construct()
@@ -18,143 +16,127 @@ class UserController extends Controller
         $this->loadView('admin/user',$data);
     }
 
-    //Standanrd User register form
+    //Standard User register
     public function register()
     {
-          //Initialize data 
-            $data =[    
-            'firstname' => '',
-            'lastname' => '',
-            'email' => '',
-            'password' => '',
-            'confirm_password' => '', 
-            'firstname_error' => '',
-            'lastname_error' => '',           
-            'email_error' => '',
-            'password_error' => '',
-            'confirm_password_error' => ''                
-        ];
-        //Load view
-        $this->loadView('admin/register',$data);
-    }
-    //Login function
-    public function login()
-    { 
-        //Avoid data send by GET method
-        if($_SERVER['REQUEST_METHOD'] == 'POST')
-        {
-             //Pocess form
-             
-            //Sanitize POST data
-            $_POST = filter_input_array(INPUT_POST, FILTER_SANITIZE_STRING);
-            //Initialize data posted
-            $email=trim($_POST['email']);
-            $pass=trim($_POST['password']);
-            //var_dump('Email: ' . $email . ' - pwd: ' . $pass); die();
-            //Initialize error message
-            $data = [
-                'email' => $email,
-                'password' => $pass,
-                'email_error' => '',
-                'password_error' => ''
-            ];
-            
-            //Validate email
-            if(empty($email))
-            {
-                $data['email_error'] = 'Veuiller saisir un email valide';
-            }
-            //Validate password
-            if(empty($pass))
-            {
-                $data['password_error'] = 'Veuiller saisir un mot de passe';
-            }
-            
-            //Check if that email exists in db valable en registration
-            if($this->userManager->findByEmail($email))
-            {               
-                //User found ok continue 
-            }
-            else
-            {
-                //User not found
-                $data['email_error'] = 'Email non trouvé!';               
-            }
-            //If all errors are empty
-            if(empty($data['email_error']) && empty($data['password_error']))
-            {             
-                //Validate
+            //Avoid data send by GET method
+            if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+                 //Pocess form
+                 
+                //Sanitize POST data
+                $_POST = filter_input_array(INPUT_POST, FILTER_SANITIZE_STRING);
+    
+                //Initialize data posted
+                $firstname=trim($_POST['firstname']);
+                $lastname=trim($_POST['lastname']);
+                $email=trim($_POST['email']);            
+                $pass=trim($_POST['password']);
+                $pass_confirm=trim($_POST['confirm_password']);
+    
+                //Initialize data
+                $data = [
+                    'firstname' => $firstname,
+                    'lastname' => $lastname,
+                    'email' => $email,
+                    'password' => $pass,                
+                    'confirm_password' => $pass_confirm,
+                    'firstname_error' => '',
+                    'lastname_error' => '',
+                    'email_error' => '',
+                    'password_error' => '',
+                    'confirm_password_error' => ''
+                ];
                 
-                //Check and set logged in User
-                $loggedInUser = $this->userManager->login($email, $pass);
-                
-                if($loggedInUser)
-                {                  
-                    //Create session
-                    $_SESSION['user_id']        = $loggedInUser->getId();
-                    $_SESSION['user_email']     = $loggedInUser->getEmail();
-                    $_SESSION['user_firstname'] = $loggedInUser->getFirstname();
-                    $_SESSION['user_lastname']  = $loggedInUser->getLastname();
-            
-                    //$_SESSION['message'] = "Vous êtes connecté";
-                    
-                    header('Location: '. URL_PATH.'homeAdmin');                    
-                    
+                //Validate firstname
+                if (empty($data['firstname'])) {
+                    $data['firstname_error'] = 'Veuiller saisir votre prénom';                    
                 }
-                else
+                //Validate lastname
+                if (empty($data['lastname'])) {
+                    $data['lastname_error'] = 'Veuiller saisir votre nom';
+                }
+                //Validate email
+                if (empty($data['email']) || !(filter_var($data['email'], FILTER_VALIDATE_EMAIL))) {
+                    $data['email_error'] = 'Veuiller saisir un email valide';
+                }
+                //Validate password
+                if (empty($data['password'])) {
+                    $data['password_error'] = 'Veuiller saisir un mot de passe';
+                } else {
+                    if (strlen($data['password']) < 6) {
+                        $data['password_error'] = 'Le mot de passe doit avoir au moins 6 caractères';
+                    }
+                }
+    
+                //Validate confirmed password
+                if (empty($data['confirm_password']))
                 {
-                    $data['password_error'] = "Mot de passe invalid";
+                    $data['confirm_password_error'] = 'Veuiller confirmer le mot de passe';
+                } else {
+                    if ($data['password'] != $data['confirm_password']) {
+                        $data['confirm_password_error'] = 'Les 2 mots de passe ne sont pas identiques';
+                    }
+                }
+                //Check if that email exists in db valable en registration
+                if ($this->userManager->findByEmail($email)) {
+                   //User found duplicate error else continue
+                   $data['email_error'] = 'Email déja utilisé !'; 
+                }
+                
+                //If all errors are empty
+                if (empty($data['firstname_error']) && empty($data['lastname_error'])
+                && empty($data['email_error']) && empty($data['password_error']) 
+                && empty($data['confirm_password_error'])) {             
+                    //Validate
+                    
+                    //Password hash
+                    $pass_hash = password_hash($pass, PASSWORD_DEFAULT);
+                    
+                    $user = new User();                       
+                    //Assign values to the new user to create
+                    $user->setFirstname($firstname)
+                            ->setLastname($lastname)
+                            ->setEmail($email)
+                            ->setPassword($pass_hash);
+                                            
+                    //insert into db using manager's create method
+                    $result = $this->userManager->create($user);
+                    header('Location: '. URL_PATH.'user/login');
+                } else {
                     //Reload view with errors
-                    $this->loadView('admin/login',$data);
-
-                }                
-            } 
-            //$data['password_error'] = "Mot de passe invalid";
-            //Reload view with errors
-            //var_dump($data);
-            //die();
-            $this->loadView('admin/login', $data);
-        }
-        else
-        {
+                    $this->loadView('admin/register',$data);
+                }            
+            } else {
             //Initialize data 
-            $data =[    
+                $data =[    
+                'firstname' => '',
+                'lastname' => '',
                 'email' => '',
-                'password' => '',            
+                'password' => '',
+                'confirm_password' => '', 
+                'firstname_error' => '',
+                'lastname_error' => '',           
                 'email_error' => '',
-                'password_error' => ''                
+                'password_error' => '',
+                'confirm_password_error' => ''                
             ];
             //Load view
-            $this->loadView('admin/login',$data);
-        }        
-    }    
-
+            $this->loadView('admin/register',$data);
+        }
+    }   
+    
     public function findById($id)
     {
         $userManager = new UserManager();
-        //on récupère l'objet correspondant à l'id specifié de la bdd dans la variable $categorie
         $user = $userManager->findById($id);
 
         require('views/updateUserView.php');
-    } 
-    //start the login form in case invalid or not supplied URL
-    public function logout()
-    {        
-        unset($_SESSION['user_id']);
-        unset($_SESSION['user_firstname']);
-        unset($_SESSION['user_lastname']);
-        unset($_SESSION['user_email']);
-        unset($_SESSION['message']);
-        session_destroy();
-        
-        header('Location: '. URL_PATH.'user/login'); 
     }
 
     public function add()
     {
         //Avoid data send by GET method
-        if ($_SERVER['REQUEST_METHOD'] == 'POST')
-        {
+        if ($_SERVER['REQUEST_METHOD'] == 'POST') {
              //Pocess form
              
             //Sanitize POST data
@@ -182,45 +164,37 @@ class UserController extends Controller
             ];
             
             //Validate firstname
-            if (empty($data['firstname']))
-            {
+            if (empty($data['firstname'])) {
                 $data['firstname_error'] = 'Veuiller saisir votre prénom';
             }
             //Validate lastname
-            if (empty($data['lastname']))
-            {
+            if (empty($data['lastname'])) {
                 $data['lastname_error'] = 'Veuiller saisir votre nom';
             }
             //Validate email
-            if (empty($data['email']) || !(filter_var($data['email'], FILTER_VALIDATE_EMAIL)))
-            {
+            if (empty($data['email']) || !(filter_var($data['email'], FILTER_VALIDATE_EMAIL))) {
                 $data['email_error'] = 'Veuiller saisir un email valide';
             }
             
             //Validate password
-            if (empty($data['password']))
-            {
+            if (empty($data['password'])) {
                 $data['password_error'] = 'Veuiller saisir un mot de passe';
             } else {
-                if (strlen($data['password']) < 6)
-                {
+                if (strlen($data['password']) < 6) {
                     $data['password_error'] = 'Le mot de passe doit avoir au moins 6 caractères';
                 }
             }
 
             //Validate password
-            if (empty($data['confirm_password']))
-            {
+            if (empty($data['confirm_password'])) {
                 $data['confirm_password_error'] = 'Veuiller confirmer le mot de passe';
             } else {
-                if ($data['password'] != $data['confirm_password'])
-                {
+                if ($data['password'] != $data['confirm_password']) {
                     $data['confirm_password_error'] = 'Les 2 mots de passe ne sont pas identiques';
                 }
             }            
-            //Check if that email exists in db valable en registration
-            if ($this->userManager->findByEmail($email))
-            {
+            //Check if that email exists in db
+            if ($this->userManager->findByEmail($email)) {
                //User found duplicate error else continue
                $data['email_error'] = 'Email déja utilisé !'; 
             }
@@ -228,8 +202,7 @@ class UserController extends Controller
             //If all errors are empty
             if (empty($data['firstname_error']) && empty($data['lastname_error'])
             && empty($data['email_error']) && empty($data['password_error']) 
-            && empty($data['confirm_password_error']))
-            {             
+            && empty($data['confirm_password_error'])) {             
                 //Validate
                 
                 //Password hash
@@ -267,14 +240,16 @@ class UserController extends Controller
             //Load view
             $this->loadView('admin/addUser',$data);
         }        
-    } 
-
+    }    
     
     public function edit($id)
-    {        
+    {
+        if (isset($_GET['token']) && ($_GET['token'] != $_SESSION['user']['token']) || empty($_GET['token'])) {
+            exit("Token périmé!");
+        }
+
         //Avoid data send by GET method
-        if ($_SERVER['REQUEST_METHOD'] == 'POST')
-        {
+        if ($_SERVER['REQUEST_METHOD'] == 'POST') {
              //Pocess form
              
             //Sanitize POST data
@@ -297,18 +272,15 @@ class UserController extends Controller
                 ];
             
             //Validate firstname
-            if (empty($data['firstname']))
-            {
+            if (empty($data['firstname'])) {
                 $data['firstname_error'] = 'Veuiller saisir votre prénom';
             }
             //Validate lastname
-            if (empty($data['lastname']))
-            {
+            if (empty($data['lastname'])) {
                 $data['lastname_error'] = 'Veuiller saisir votre nom';
             }
             //Validate email
-            if (empty($data['email']) || !(filter_var($data['email'], FILTER_VALIDATE_EMAIL)))
-            {
+            if (empty($data['email']) || !(filter_var($data['email'], FILTER_VALIDATE_EMAIL))) {
                 $data['email_error'] = 'Veuiller saisir un email valide';
             }            
             
@@ -323,8 +295,7 @@ class UserController extends Controller
             
             //If all errors are empty
             if (empty($data['firstname_error']) && empty($data['lastname_error'])
-            && empty($data['email_error']))
-            {             
+            && empty($data['email_error'])) {             
                 //Validate
                 
                 //Get user to update from Manager
@@ -332,8 +303,9 @@ class UserController extends Controller
                 //Assign values to the new user to create
                 $userToUpdate->setFirstname($data['firstname'])
                             ->setLastname($data['lastname'])
-                            ->setEmail($data['email']);
-                                     
+                            ->setEmail($data['email'])
+                            ->setUpdatedAt(date("Y-m-d H:i:s"));
+                                 
                 //insert into db using manager's create method
                 $this->userManager->update($userToUpdate);                
             
@@ -349,7 +321,7 @@ class UserController extends Controller
             //Check for ownership
             if ($user->getId() != $_SESSION['user_id']){
                 //TO BE USED FOR posts
-                header('Location: '. URL_PATH.'user/index');                    
+                header('Location: '. URL_PATH.'user/index');
             }
             //Initialize data for edit form
             $data = [
@@ -358,8 +330,8 @@ class UserController extends Controller
                 'lastname' => $user->getlastname(),
                 'email' => $user->getEmail(),
                 'firstname_error' => '',
-                'lastname_error' => '',            
-                'email_error' => ''                
+                'lastname_error' => '', 
+                'email_error' => '' 
             ];
             //Load view
             $this->loadView('admin/editUser',$data);
@@ -368,23 +340,20 @@ class UserController extends Controller
 
     public function delete($id)
     {
-        //on fait appel au manager
-        $userManager = new UserManager();
-
-        /*on récupère l'objet correspondant à l'id specifié de la bdd.
-        On stocke cet objet dans la variable $categorie */
-        $user = $userManager->findById($id);
-
-        //On supprime alors cet objet en appelant la méthode du manager et en lui passant l'objet en question
-        $isDeleteOk = $userManager->delete($user);
-
-        if (!$isDeleteOk)
-        {
-            die('Impossible de supprimer l\'élémént!');
-
-        } else {
-            header('Location: index.php?action=listUsers');
+        if (isset($_GET['token']) && ($_GET['token'] != $_SESSION['user']['token']) || empty($_GET['token'])) {
+            exit("Token périmé!");
         }
 
+        $userManager = new UserManager();
+
+        $user = $userManager->findById($id);
+
+        $isDeleteOk = $userManager->delete($user);
+
+        if (!$isDeleteOk) {
+            die('Impossible de supprimer l\'élémént!');
+        } else {
+            header('Location: '. URL_PATH.'user/index');
+        }
     }
 }
